Watch for PulseAudio streams on module-combine-sink sinks
We control starting and stopping of playback on devices by
observing the number of streams playing on the device.
If there is a PulseAudio module-combine-sink which has one
of our sinks as a slave, we should count the streams that are
playing on the combine sink when deciding whether to start or
stop playback on our sink. To do this, we simulate that any
stream that is playing on a combine sync is actually playing
on each of its slaves instead.